last commit

+++ back/src/main/java/com/example/bureau/security/jwt/JwtUtils.java
@@ -24,7 +24,7 @@ public class JwtUtils {
     private String jwtSecret;
 
     @Value("${yess.app.jwtExpirationMs}")
-    private int jwtExpirationMs;
+    private Long jwtExpirationMs;
 
    /* public String generateJwtToken(Authentication authentication) {
